+++ cli/main.js
